requirements.txt
flask==3.0.2
python-dotenv==1.0.1
unidecode==1.3.8
scikit-learn==1.6.1
numpy==1.26.4
rapidfuzz==3.12.1
jarowinkler==2.0.1
pandas==2.2.2